# openapi/nova-browser-api.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "Body_register_upload_api_v1_source_datasets_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "manifest": {
            "title": "Manifest",
            "type": "string"
          }
        },
        "required": [
          "manifest",
          "file"
        ],
        "title": "Body_register_upload_api_v1_source_datasets_post",
        "type": "object"
      },
...
      "Body_validate_upload_api_v1_source_datasets_validate_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "manifest": {
            "title": "Manifest",
            "type": "string"
          }
        },
        "required": [
          "manifest",
          "file"
        ],
        "title": "Body_validate_upload_api_v1_source_datasets_validate_post",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "detail": {
            "title": "Detail"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "required": [
          "detail",
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "SourceDatasetRegistered": {
        "additionalProperties": false,
        "properties": {
          "analysis_ready": {
            "const": false,
            "default": false,
            "title": "Analysis Ready",
            "type": "boolean"
          },
          "bounds": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bounds"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "dataset_id": {
            "format": "uuid",
            "title": "Dataset Id",
            "type": "string"
          },
          "format": {
            "title": "Format",
            "type": "string"
          },
          "record_count": {
            "title": "Record Count",
            "type": "integer"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "structural_validation_only": {
            "const": true,
            "default": true,
            "title": "Structural Validation Only",
            "type": "boolean"
          }
        },
        "required": [
          "category",
          "format",
          "record_count",
          "bounds",
          "sha256",
          "dataset_id"
        ],
        "title": "SourceDatasetRegistered",
        "type": "object"
      },
      "SourceDatasetValidation": {
        "additionalProperties": false,
        "properties": {
          "analysis_ready": {
            "const": false,
            "default": false,
            "title": "Analysis Ready",
            "type": "boolean"
          },
          "bounds": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bounds"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "format": {
            "title": "Format",
            "type": "string"
          },
          "record_count": {
            "title": "Record Count",
            "type": "integer"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "structural_validation_only": {
            "const": true,
            "default": true,
            "title": "Structural Validation Only",
            "type": "boolean"
          }
        },
        "required": [
          "category",
          "format",
          "record_count",
          "bounds",
          "sha256"
        ],
        "title": "SourceDatasetValidation",
        "type": "object"
      },
      "SourceReadinessReview": {
        "additionalProperties": false,
        "properties": {
          "annual_record_complete_verified": {
            "default": false,
            "title": "Annual Record Complete Verified",
            "type": "boolean"
          },
          "coverage_complete_verified": {
            "default": false,
            "title": "Coverage Complete Verified",
            "type": "boolean"
          },
          "crs_datum_verified": {
            "const": true,
            "title": "Crs Datum Verified",
            "type": "boolean"
          },
          "dem_conditioning_verified": {
            "default": false,
            "title": "Dem Conditioning Verified",
            "type": "boolean"
          },
          "evidence_refs": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Evidence Refs",
            "type": "array"
          },
          "hydraulic_model_verified": {
            "default": false,
            "title": "Hydraulic Model Verified",
            "type": "boolean"
          },
          "hydrologic_coverage_verified": {
            "default": false,
            "title": "Hydrologic Coverage Verified",
            "type": "boolean"
          },
          "licence_verified": {
            "const": true,
            "title": "Licence Verified",
            "type": "boolean"
          },
          "provenance_verified": {
            "const": true,
            "title": "Provenance Verified",
            "type": "boolean"
          },
          "qa_verified": {
            "const": true,
            "title": "Qa Verified",
            "type": "boolean"
          },
          "temporal_coverage_verified": {
            "const": true,
            "title": "Temporal Coverage Verified",
            "type": "boolean"
          },
          "uncertainty_reviewed": {
            "const": true,
            "title": "Uncertainty Reviewed",
            "type": "boolean"
          }
        },
        "required": [
          "evidence_refs",
          "qa_verified",
          "licence_verified",
          "provenance_verified",
          "crs_datum_verified",
          "temporal_coverage_verified",
          "uncertainty_reviewed"
        ],
        "title": "SourceReadinessReview",
        "type": "object"
      },
      "SourceReadinessStatus": {
        "additionalProperties": false,
        "properties": {
          "analysis_ready": {
            "title": "Analysis Ready",
            "type": "boolean"
          },
          "dataset_id": {
            "format": "uuid",
            "title": "Dataset Id",
            "type": "string"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          }
        },
        "required": [
          "dataset_id",
          "analysis_ready",
          "sha256"
        ],
        "title": "SourceReadinessStatus",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "NOVA-GeoRisk Browser API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/source-datasets": {
      "post": {
        "operationId": "register_upload_api_v1_source_datasets_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_register_upload_api_v1_source_datasets_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDatasetRegistered"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authorization failed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Register Upload",
        "tags": [
          "Source datasets"
        ]
      }
    },
    "/api/v1/source-datasets/catalogue": {
      "get": {
        "operationId": "catalogue_api_v1_source_datasets_catalogue_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Catalogue Api V1 Source Datasets Catalogue Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authorization failed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Catalogue",
        "tags": [
          "Source datasets"
        ]
      }
    },
    "/api/v1/source-datasets/validate": {
      "post": {
        "operationId": "validate_upload_api_v1_source_datasets_validate_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_validate_upload_api_v1_source_datasets_validate_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDatasetValidation"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authorization failed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Validate Upload",
        "tags": [
          "Source datasets"
        ]
      }
    },
    "/api/v1/source-datasets/{dataset_id}/approve": {
      "post": {
        "operationId": "approve_source_dataset_api_v1_source_datasets__dataset_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceReadinessReview"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceReadinessStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authorization failed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Approve Source Dataset",
        "tags": [
          "Source datasets"
        ]
      }
    },
    "/api/v1/source-datasets/{dataset_id}/revoke": {
      "post": {
        "operationId": "revoke_source_dataset_api_v1_source_datasets__dataset_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceReadinessStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authorization failed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Revoke Source Dataset",
        "tags": [
          "Source datasets"
        ]
      }
    },
...
  }
}